图片上传问题

+++ living/living/Model/Request/HomePage/LMHomelistequest.m
@@ -16,10 +16,10 @@
     if (self) {
         NSMutableDictionary *body = [NSMutableDictionary new];
         if (pageIndex != -1) {
-            [body setObject:[NSString stringWithFormat:@"%ld", pageIndex] forKey:@"pageIndex"];
+            [body setObject:[NSString stringWithFormat:@"%ld", (long)pageIndex] forKey:@"pageIndex"];
         }
         if (pageSize != -1) {
-            [body setObject:[NSString stringWithFormat:@"%ld", pageSize] forKey:@"pageSize"];
+            [body setObject:[NSString stringWithFormat:@"%ld", (long)pageSize] forKey:@"pageSize"];
         }
         
         NSMutableDictionary *parmDic = [self params];
